working on CNN ... logging

+++ elixer/hetdex.py
@@ -14471,7 +14471,7 @@ class HETDEX:
                             #since only running one, the detectid (entry_id) does not really matter
                             #this needs to be wrapped in cnn module
                             if ML_CNN is not None:
-                                cnn_t = ML_CNN.process_detections([datakeep['detobj'].ml_2d_fiber_sum], [datakeep['detobj'].entry_id])
+                                cnn_t = ML_CNN.process_detections([datakeep['detobj'].ml_2d_fiber_sum.astype(np.float32)], [datakeep['detobj'].entry_id])
                                 if cnn_t is not None:
                                     datakeep['detobj'].ml_cnn_score = cnn_t[0]['CNN_Score_2D_Spectra']
                                     log.info(f"eid({datakeep['detobj'].entry_id}) CNN score = {datakeep['detobj'].ml_cnn_score}")
